cleaned up the code and made the files sprated
