+++ src/getRecipes.js
@@ -1,17 +1,8 @@
-// importing functions
+// import functions
 import { getData } from './getData.js'
 
-// selecting elements
-const searchBtn = document.getElementById('search-Btn')
-const searchInput = document.getElementById('search-Input')
-
-// gelobal variables
-let searchTerm = `pizza`
-
-// functions
-
 // fter click on seach button the function will be called
-const getRecipes = (searchTerm) => {
+export const getRecipes = (searchTerm) => {
   console.log('passed', searchInput.value)
   let seacrchTimer = 0
   //
@@ -21,21 +12,8 @@ const getRecipes = (searchTerm) => {
 
     if (seacrchTimer > 4) {
       seacrchTimer = 0
-      searchTerm = searchInput.value
       getData(searchTerm)
       clearInterval(seachInterval)
     }
   }, 500)
 }
-
-// events
-
-// console.log('Hello World!')
-searchBtn.addEventListener('click', () => {
-  searchTerm = searchInput.value
-  searchTerm
-    ? getRecipes(searchTerm)
-    : console.log('Please enter a search term')
-})
-
-getData(searchTerm)
