+++ src/index.html
@@ -46,6 +46,7 @@
     </section>
     <script type="module" src="getData.js" defer></script>
     <script type="module" src="showData.js" defer></script>
+    <script type="module" src="getRecipesHandler.js" defer></script>
     <script type="module" src="getRecipes.js"></script>
   </body>
 </html>
